Move clowder yaml components into model submodule
