+++ test/unittests/test_fork.py
@@ -3,9 +3,10 @@
 import os
 import sys
 import unittest
+
+from clowder.model.fork import Fork
+from clowder.model.source import Source
 from test.unittests.shared import GITHUB_SSH_SOURCE_YAML
-from clowder.fork import Fork
-from clowder.source import Source
 
 
 class ForkTest(unittest.TestCase):
