+++ test/unittests/test_source.py
@@ -2,11 +2,12 @@
 
 import sys
 import unittest
+
+from clowder.model.source import Source
 from test.unittests.shared import (
     GITHUB_HTTPS_SOURCE_YAML,
     GITHUB_SSH_SOURCE_YAML
 )
-from clowder.source import Source
 
 
 class SourceTest(unittest.TestCase):
